requirements.txt
flask
openai
python-dotenv
pdfplumber
gunicorn